* Simplify spell affect implementation * Fixed: calculate correctly spell mod value sent to client (discern modifier type: flat/pct)
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@6322 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Unit.h
@@ -661,7 +661,7 @@ class MANGOS_DLL_SPEC Unit : public WorldObject
 {
     public:
         typedef std::set<Unit*> AttackerSet;
-        typedef std::pair<uint32, uint32> spellEffectPair;
+        typedef std::pair<uint32, uint8> spellEffectPair;
         typedef std::multimap< spellEffectPair, Aura*> AuraMap;
         typedef std::list<Aura *> AuraList;
         typedef std::list<DiminishingReturn> Diminishing;
